Entities refactored. Primitive types changed to wrapper types

+++ Homeworks/Homework-4/turkcellmarket/src/main/java/com/example/turkcellmarket/services/dtos/product/requests/ProductAddRequest.java
@@ -23,8 +23,8 @@ public class ProductAddRequest {
     private String description;
 
     @Min(1)
-    private double price;
+    private Double price;
 
     @Min(1)
-    private double stock;
+    private Integer stock;
 }
